Move IndexedNestedRingTester.h to include
