+++ include/geos/operation/valid/IndexedNestedRingTester.h
